{"schema":"decodex/commit/1","summary":"Modularize control plane and lane projection boundaries","authority":"manual"}
